refactor: move validation handler out of webhooks package (#4556)
